+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 ---------
 
+master
+~~~~~~
+
+a) [FEATURE] Add task "file:copy_dirs_ignore_existing" which copy directories from previous release except for those
+    folders which already exists in that folder.
+
 14.0.1
 ~~~~~~~